LayoutTests/platform/ios-sim/storage/indexeddb/transaction-readwrite-exclusive-expected.txt

CONSOLE MESSAGE: line 14: TypeError: undefined is not an object (evaluating 'request.onerror = unexpectedBlockedCallback')
Check that readwrite transactions with overlapping scopes do not run in parallel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "transaction-readwrite-exclusive.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onerror = unexpectedBlockedCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
